chore(Report): Refractored scorecard specific pdf generation.
