+++ src/krisi/report/pdf.py
@@ -1,22 +1,26 @@
 import base64
 import pkgutil
-from typing import List, Optional
+from typing import TYPE_CHECKING, Dict, List
 
 import plotly.graph_objects as go
 
-from krisi.report.type import PathConst
+from krisi.report.type import InteractiveFigure, PathConst
+from krisi.utils.iterable_helpers import remove_nans
+
+if TYPE_CHECKING:
+    from krisi.evaluate.metric import Metric
 
 
-def figure_to_html(figure: go.Figure) -> str:
+def __figure_to_html(figure: go.Figure) -> str:
     image = str(base64.b64encode(figure.to_image(format="png", scale=5)))[2:-1]
     return f'<img style="max-width:100%; max-height:100%;" src="data:image/png;base64,{image}"/>'
 
 
-def convert_figures(figures: List[go.Figure]) -> str:
-    return "<br>".join([figure_to_html(figure) for figure in figures])
+def convert_figures_to_html(figures: List[go.Figure]) -> str:
+    return "<br>".join([__figure_to_html(figure) for figure in figures])
 
 
-def create_html_report(
+def __create_html_report(
     template_file: str, html_elements_to_inject: dict[str, str]
 ) -> str:
     template_html = pkgutil.get_data(__name__, template_file)
@@ -27,7 +31,7 @@ def create_html_report(
     return template_html
 
 
-def convert_html_to_pdf(
+def __convert_html_to_pdf(
     source_html: str, output_path: str, report_name: str, css_template_url: str
 ) -> None:
     import os
@@ -54,5 +58,43 @@ def create_pdf_report(
     html_elements_to_inject: dict[str, str] = dict(),
 ):
     html_elements_to_inject["title"] = title
-    report_html = create_html_report(html_template_url, html_elements_to_inject)
-    convert_html_to_pdf(report_html, path, f"{title}.pdf", css_template_url)
+    report_html = __create_html_report(html_template_url, html_elements_to_inject)
+    __convert_html_to_pdf(report_html, path, f"{title}.pdf", css_template_url)
+
+
+def get_waterfall_metric_html(metrics: List["Metric"]) -> str:
+    custom_metric_interactive_diagrams = remove_nans(
+        [metric.get_diagrams() for metric in metrics]
+    )
+    custom_diagrams = [
+        plot_func
+        for plot_funcs in custom_metric_interactive_diagrams
+        for plot_func in plot_funcs
+    ]
+
+    html_images = convert_figures_to_html(
+        [
+            interactive_diagram.get_figure(width=900.0)
+            for interactive_diagram in custom_diagrams
+        ]
+    )
+
+    return html_images
+
+
+def append_sizes(
+    diagram_dict: Dict[str, InteractiveFigure]
+) -> Dict[str, InteractiveFigure]:
+
+    size_dict = dict(
+        residuals_display_acf_plot=(750.0, 750.0),
+        residuals_display_density_plot=(350.0, 350.0),
+        residuals_display_time_series=(1500.0, 750.0),
+    )
+
+    for key, value in size_dict.items():
+        if key in diagram_dict.keys():
+            diagram_dict[key].width = value[0]
+            diagram_dict[key].height = value[1]
+
+    return diagram_dict
